LLM server/EmbeddingProcess/TextExtraction/ImageOCR_Server/requirements.txt
torch==2.0.1
transformers>=4.38.2
timm>=0.9.16
accelerate
sentencepiece
attrdict
einops
flask
flask-cors
pillow